+++ src/dockerHighlight.ts
@@ -13,13 +13,13 @@ import { Util } from './docker';
 
 export class DockerHighlight {
 
-    public computeHighlightRanges(textDocument: TextDocumentIdentifier, content: string, position: Position): DocumentHighlight[] {
+    public computeHighlightRanges(content: string, position: Position): DocumentHighlight[] {
         let dockerfile = DockerfileParser.parse(content);
         let provider = new DockerDefinition();
-        let location = provider.computeDefinition(textDocument, content, position);
-        let image = location === null ? dockerfile.getContainingImage(position) : dockerfile.getContainingImage(location.range.start);
+        const definitionRange = provider.computeDefinitionRange(content, position);
+        let image = definitionRange === null ? dockerfile.getContainingImage(position) : dockerfile.getContainingImage(definitionRange.start);
         const highlights: DocumentHighlight[] = [];
-        if (location === null) {
+        if (definitionRange === null) {
             for (let instruction of dockerfile.getCOPYs()) {
                 let flag = instruction.getFromFlag();
                 if (flag) {
@@ -73,12 +73,12 @@ export class DockerHighlight {
                 }
             }
         } else {
-            let document = TextDocument.create(textDocument.uri, "", 0, content);
-            let definition = document.getText().substring(document.offsetAt(location.range.start), document.offsetAt(location.range.end));
+            let document = TextDocument.create("", "", 0, content);
+            let definition = document.getText().substring(document.offsetAt(definitionRange.start), document.offsetAt(definitionRange.end));
             for (let from of dockerfile.getFROMs()) {
                 let range = from.getBuildStageRange();
-                if (range && range.start.line === location.range.start.line) {
-                    highlights.push(DocumentHighlight.create(location.range, DocumentHighlightKind.Write));
+                if (range && range.start.line === definitionRange.start.line) {
+                    highlights.push(DocumentHighlight.create(definitionRange, DocumentHighlightKind.Write));
                     for (let instruction of dockerfile.getCOPYs()) {
                         let flag = instruction.getFromFlag();
                         if (flag) {
@@ -120,7 +120,7 @@ export class DockerHighlight {
 
             for (const arg of dockerfile.getInitialARGs()) {
                 const property = arg.getProperty();
-                if (property && Util.rangeEquals(property.getNameRange(), location.range)) {
+                if (property && Util.rangeEquals(property.getNameRange(), definitionRange)) {
                     for (const from of dockerfile.getFROMs()) {
                         for (const variable of from.getVariables()) {
                             if (variable.getName() === definition) {