+++ src/dockerRename.ts
@@ -12,7 +12,7 @@ export class DockerRename {
     public rename(textDocument: TextDocumentIdentifier, content: string, position: Position, newName: string): TextEdit[] {
         const edits: TextEdit[] = [];
         const highlighter = new DockerHighlight();
-        const highlightRanges = highlighter.computeHighlightRanges(textDocument, content, position);
+        const highlightRanges = highlighter.computeHighlightRanges(content, position);
         for (const highlightRange of highlightRanges) {
             edits.push(TextEdit.replace(highlightRange.range, newName));
         }